services/fits-processor/requirements.txt
grpcio==1.76.0
grpcio-tools==1.76.0
astropy==6.1.2
matplotlib==3.10.0
numpy>=2.0.0
protobuf==6.33.4